atom_chat/requirements.txt
Django==5.1.3
daphne
channels
channels-redis
redis
psycopg2-binary
djangorestframework
djangorestframework-simplejwt